Refactor imports for @hedvig-ui/animations
