+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { css, Global, ThemeProvider } from '@emotion/react'
 import styled from '@emotion/styled'
+import { StandaloneMessage } from '@hedvig-ui'
 import { colorsV3, fonts, getCdnFontFaces } from '@hedviginsurance/brand'
 import { CssBaseline } from '@material-ui/core'
 import { MuiThemeProvider } from '@material-ui/core/styles'
@@ -8,7 +9,6 @@ import { DashboardPage } from 'components/dashboard'
 import { QuestionsPage } from 'components/questions'
 import Breadcrumbs from 'components/shared/navigation/breadcrumbs/Breadcrumbs'
 import { VerticalMenu } from 'components/shared/navigation/sidebar/VerticalMenu'
-import { StandaloneMessage } from 'hedvig-ui/animations/standalone-message'
 import {
   darkTheme,
   darkUiTheme,
