+++ src/components/claims/index.tsx
@@ -1,10 +1,9 @@
-import { Spacing } from '@hedvig-ui'
+import { LoadingMessage, Spacing } from '@hedvig-ui'
 import { Claim } from 'api/generated/graphql'
 import { ListPage } from 'components/shared'
 import { Paginator } from 'components/shared/paginator/Paginator'
 import { useListClaims } from 'graphql/use-list-claims'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import { LoadingMessage } from 'hedvig-ui/animations/standalone-message'
 import React, { useEffect } from 'react'
 import { RouteComponentProps, useHistory } from 'react-router'
 import { Header } from 'semantic-ui-react'
