+++ src/components/member/tabs/contracts-tab/agreement/CreateQuoteFromAgreement.tsx
@@ -1,12 +1,11 @@
 import styled from '@emotion/styled'
-import { Button, ThirdLevelHeadline } from '@hedvig-ui'
+import { Button, StandaloneMessage, ThirdLevelHeadline } from '@hedvig-ui'
 import { Contract, GenericAgreement } from 'api/generated/graphql'
 import {
   createQuoteFromAgreementOptions,
   useCreateQuoteFromAgreement,
 } from 'graphql/use-create-quote'
 import { useQuotes } from 'graphql/use-get-quotes'
-import { StandaloneMessage } from 'hedvig-ui/animations/standalone-message'
 import React from 'react'
 import { toast } from 'react-hot-toast'
 import { isExpired } from 'utils/quote'
