+++ src/components/member/tabs/debt-tab/index.tsx
@@ -5,17 +5,15 @@ import {
   InfoContainer,
   InfoRow,
   InfoText,
+  LoadingMessage,
   MainHeadline,
   Spacing,
+  StandaloneMessage,
 } from '@hedvig-ui'
 import { WhitelistMemberButton } from 'components/member/tabs/debt-tab/WhitelistMemberButton'
 import { RefreshButton } from 'components/member/tabs/shared/refresh-button'
 import { useGetPerson } from 'graphql/use-get-person'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import {
-  LoadingMessage,
-  StandaloneMessage,
-} from 'hedvig-ui/animations/standalone-message'
 import React from 'react'
 import { ArrowRepeat } from 'react-bootstrap-icons'
 import { Market } from 'types/enums'
