+++ src/components/member/tabs/quote-tab/index.tsx
@@ -1,11 +1,11 @@
-import { getTextFromEnumValue } from '@hedvig-ui'
+import {
+  getTextFromEnumValue,
+  LoadingMessage,
+  StandaloneMessage,
+} from '@hedvig-ui'
 import { Quote } from 'api/generated/graphql'
 import { useQuotes } from 'graphql/use-get-quotes'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import {
-  LoadingMessage,
-  StandaloneMessage,
-} from 'hedvig-ui/animations/standalone-message'
 import React from 'react'
 import { Tab } from 'semantic-ui-react'
 import { ContractType, Market, QuoteProductType } from 'types/enums'
