+++ src/components/questions/index.tsx
@@ -1,13 +1,14 @@
-import { Spacing, ThirdLevelHeadline } from '@hedvig-ui'
+import {
+  LoadingMessage,
+  Spacing,
+  StandaloneMessage,
+  ThirdLevelHeadline,
+} from '@hedvig-ui'
 import { NumberMemberGroupsRadioButtons } from 'components/questions/number-member-groups-radio-buttons'
 import { QuestionGroups } from 'components/questions/questions-list/QuestionGroups'
 import { ListPage } from 'components/shared'
 import { useQuestionGroups } from 'graphql/use-question-groups'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import {
-  LoadingMessage,
-  StandaloneMessage,
-} from 'hedvig-ui/animations/standalone-message'
 import React from 'react'
 import { useInsecurePersistentState } from 'utils/state'
 import { FilterState, QuestionsFilter } from './filter'
